requirements.txt
accelerate==0.25.0
aiohttp==3.9.3
aiosignal==1.3.1
alembic==1.13.1
annotated-types==0.6.0
anyio==4.3.0
appdirs==1.4.4
arxiv==2.1.0
asgiref==3.7.2
asttokens==2.4.1
attrs==23.2.0
autogenstudio==0.0.47
backoff==2.2.1
bcrypt==4.1.2
beautifulsoup4==4.12.3
bitsandbytes==0.41.3.post2
Brotli==1.1.0
build==1.1.1
CacheControl==0.13.1
cachetools==5.3.3
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
chroma-hnswlib==0.7.3
chromadb==0.4.24
cleo==2.1.0
click @ file:///private/tmp/python-click-20231113-5463-1enrr1r/click-8.1.7
coloredlogs==15.0.1
contextlib2==21.6.0
contourpy==1.2.0
crashtest==0.4.1
crewai==0.22.5
crewai-tools==0.0.15
cryptography==42.0.5
curl-cffi==0.5.10
cycler==0.12.1
dataclasses-json==0.6.4
datasets==2.16.1
decorator==5.1.1
Deprecated==1.2.14
deprecation==2.1.0
dill==0.3.7
diskcache==5.6.3
distlib==0.3.8
distro==1.9.0
docker==6.1.3
docstring-parser==0.15
duckduckgo_search==4.1.1
dulwich==0.21.7
embedchain==0.1.94
executing==2.0.1
fastapi==0.110.0
fastjsonschema==2.19.0
feedparser==6.0.10
filelock==3.13.1
FLAML==2.1.1
flatbuffers==24.3.7
fonttools==4.46.0
frozenlist==1.4.1
fsspec==2024.2.0
gitdb==4.0.11
GitPython==3.1.42
google-api-core==2.17.1
google-auth==2.28.2
google-cloud-aiplatform==1.43.0
google-cloud-bigquery==3.19.0
google-cloud-core==2.4.1
google-cloud-resource-manager==1.12.3
google-cloud-storage==2.15.0
google-crc32c==1.5.0
google-resumable-media==2.7.0
googleapis-common-protos==1.63.0
gptcache==0.1.43
grpc-google-iam-v1==0.13.0
grpcio==1.62.1
grpcio-status==1.62.1
h11==0.14.0
httpcore==1.0.4
httptools==0.6.1
httpx==0.27.0
huggingface-hub==0.21.4
humanfriendly==10.0
idna==3.6
importlib-metadata==6.11.0
importlib_resources==6.1.3
iniconfig==2.0.0
installer==0.7.0
instructor==0.5.2
ipython==8.22.2
jaraco.classes==3.3.0
jedi==0.19.1
Jinja2==3.1.2
joblib==1.3.2
jsonpatch==1.33
jsonpointer==2.4
keyring==24.3.0
kiwisolver==1.4.5
kubernetes==29.0.0
lancedb==0.5.7
langchain==0.1.11
langchain-community==0.0.27
langchain-core==0.1.30
langchain-experimental==0.0.47
langchain-openai==0.0.5
langchain-text-splitters==0.0.1
langsmith==0.1.23
litellm==1.12.0
llama_cpp_python==0.2.27
lxml==5.0.1
Mako==1.3.2
markdown-it-py==3.0.0
MarkupSafe==2.1.5
marshmallow==3.21.1
matplotlib==3.8.2
matplotlib-inline==0.1.6
mdurl==0.1.2
mmh3==4.1.0
monotonic==1.6
more-itertools==10.1.0
mpmath==1.3.0
msgpack==1.0.7
multidict==6.0.5
multiprocess==0.70.15
mutagen==1.47.0
mypy-extensions==1.0.0
networkx==3.2.1
nltk==3.8.1
nodeenv==1.8.0
numexpr==2.8.8
numpy==1.26.4
oauthlib==3.2.2
onnxruntime==1.17.1
openai==1.13.3
opentelemetry-api==1.23.0
opentelemetry-exporter-otlp-proto-common==1.23.0
opentelemetry-exporter-otlp-proto-grpc==1.23.0
opentelemetry-exporter-otlp-proto-http==1.23.0
opentelemetry-instrumentation==0.44b0
opentelemetry-instrumentation-asgi==0.44b0
opentelemetry-instrumentation-fastapi==0.44b0
opentelemetry-proto==1.23.0
opentelemetry-sdk==1.23.0
opentelemetry-semantic-conventions==0.44b0
opentelemetry-util-http==0.44b0
optimum==1.16.1
orjson==3.9.15
outcome==1.3.0.post0
overrides==7.7.0
packaging==23.2
pandas==2.1.3
parso==0.8.3
pexpect==4.9.0
Pillow==10.1.0
pkginfo==1.9.6
platformdirs==3.11.0
pluggy==1.4.0
poetry==1.7.1
poetry-core==1.8.1
poetry-plugin-export==1.6.0
posthog==3.5.0
prompt-toolkit==3.0.43
proto-plus==1.23.0
protobuf==4.25.3
psutil==5.9.7
psycopg==3.1.18
psycopg-binary==3.1.18
psycopg-pool==3.2.1
ptyprocess==0.7.0
pulsar-client==3.4.0
pure-eval==0.2.2
py==1.11.0
pyarrow==15.0.1
pyarrow-hotfix==0.6
pyasn1==0.5.1
pyasn1-modules==0.3.0
pyautogen==0.2.18
pycparser==2.21
pycryptodomex==3.20.0
pydantic==2.6.3
pydantic_core==2.16.3
pygame==2.5.2
PyGithub==1.59.1
Pygments @ file:///private/tmp/pygments-20231121-5274-1rlyumb/pygments-2.17.2
PyJWT==2.8.0
pylance==0.9.18
PyNaCl==1.5.0
pyparsing==3.1.1
pypdf==3.17.4
PyPika==0.48.9
pyproject_hooks==1.0.0
pyright==1.1.353
pysbd==0.3.4
PySocks==1.7.1
pytest==8.1.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.0
pytube==15.0.0
pytz==2023.3.post1
PyYAML==6.0.1
rapidfuzz==3.6.0
ratelimiter==1.2.0.post0
redis==5.0.1
regex==2023.12.25
requests==2.31.0
requests-oauthlib==1.4.0
requests-toolbelt==1.0.0
retry==0.9.2
rich==13.7.1
rq==1.15.1
rsa==4.9
safetensors==0.4.1
schema==0.7.5
scikit-learn==1.3.2
scipy==1.11.4
selenium==4.18.1
semver==3.0.2
sentence-transformers==2.2.2
sentencepiece==0.1.99
sgmllib3k==1.0.0
shapely==2.0.3
shellingham==1.5.4
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
sortedcontainers==2.4.0
soundfile==0.12.1
soupsieve==2.5
SQLAlchemy==2.0.28
stack-data==0.6.3
starlette==0.36.3
sympy==1.12
tabulate @ file:///private/tmp/python-tabulate-20231113-17262-1dzbtg2/tabulate-0.9.0
tenacity==8.2.3
termcolor==2.4.0
threadpoolctl==3.2.0
tiktoken==0.5.2
tokenizers==0.15.2
tomlkit==0.12.3
torch==2.1.2
torchaudio==2.1.2
torchvision==0.16.2
tqdm==4.66.2
traitlets==5.14.1
transformers @ git+https://github.com/huggingface/transformers@7226f3d2b06316a1623b7b557cc32f360a854860
trio==0.24.0
trio-websocket==0.11.1
trove-classifiers==2023.11.29
typer==0.9.0
typing-inspect==0.9.0
typing_extensions==4.10.0
tzdata==2023.3
urllib3==2.2.1
uvicorn==0.28.0
uvloop==0.19.0
virtualenv==20.25.0
watchfiles==0.21.0
wcwidth==0.2.13
websocket-client==1.7.0
websockets==12.0
wikipedia==1.4.0
wrapt==1.16.0
wsproto==1.2.0
xattr==0.10.1
xxhash==3.4.1
yarl==1.9.4
youtube-transcript-api==0.6.2
yt-dlp==2023.12.30
zipp==3.17.0
